requirements.txt
fastapi
uvicorn
sqlmodel
pydantic-settings
sqlalchemy